+++ src/unstructured_client/models/operations/partition.py
@@ -21,15 +21,11 @@ from unstructured_client.utils import FieldMetadata, HeaderMetadata, RequestMeta
 PARTITION_SERVER_SAAS_API = "saas-api"
 r"""Serverless SaaS API"""
 
-PARTITION_SERVER_FREE_API = "free-api"
-r"""Hosted API Free"""
-
 PARTITION_SERVER_DEVELOPMENT = "development"
 r"""Development server"""
 
 PARTITION_SERVERS = {
     PARTITION_SERVER_SAAS_API: "https://api.unstructuredapp.io",
-    PARTITION_SERVER_FREE_API: "https://api.unstructured.io",
     PARTITION_SERVER_DEVELOPMENT: "http://localhost:8000",
 }
 
@@ -61,7 +57,7 @@ class PartitionRequest(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)