+++ src/unstructured_client/models/shared/databricksvdtdestinationconnectorconfig.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 import pydantic
 from pydantic import model_serializer
+from typing import Optional
 from typing_extensions import Annotated, NotRequired, TypedDict
 from unstructured_client.types import (
     BaseModel,
@@ -15,37 +16,66 @@ from unstructured_client.types import (
 
 class DatabricksVDTDestinationConnectorConfigTypedDict(TypedDict):
     catalog: str
-    database: str
-    table_name: str
+    http_path: str
+    server_hostname: str
     volume: str
-    schema_: NotRequired[Nullable[str]]
+    client_id: NotRequired[Nullable[str]]
+    client_secret: NotRequired[Nullable[str]]
+    database: NotRequired[str]
+    schema_: NotRequired[str]
+    table_name: NotRequired[Nullable[str]]
+    token: NotRequired[Nullable[str]]
     volume_path: NotRequired[Nullable[str]]
 
 
 class DatabricksVDTDestinationConnectorConfig(BaseModel):
     catalog: str
 
-    database: str
+    http_path: str
 
-    table_name: str
+    server_hostname: str
 
     volume: str
 
-    schema_: Annotated[OptionalNullable[str], pydantic.Field(alias="schema")] = UNSET
+    client_id: OptionalNullable[str] = UNSET
+
+    client_secret: OptionalNullable[str] = UNSET
+
+    database: Optional[str] = "default"
+
+    schema_: Annotated[Optional[str], pydantic.Field(alias="schema")] = "default"
+
+    table_name: OptionalNullable[str] = UNSET
+
+    token: OptionalNullable[str] = UNSET
 
     volume_path: OptionalNullable[str] = UNSET
 
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
-        optional_fields = ["schema", "volume_path"]
-        nullable_fields = ["schema", "volume_path"]
+        optional_fields = [
+            "client_id",
+            "client_secret",
+            "database",
+            "schema",
+            "table_name",
+            "token",
+            "volume_path",
+        ]
+        nullable_fields = [
+            "client_id",
+            "client_secret",
+            "table_name",
+            "token",
+            "volume_path",
+        ]
         null_default_fields = []
 
         serialized = handler(self)
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)