+++ src/unstructured_client/models/shared/jobinformation.py
@@ -2,19 +2,31 @@
 
 from __future__ import annotations
 from .jobstatus import JobStatus
+from .nodefilemetadata import NodeFileMetadata, NodeFileMetadataTypedDict
+from .workflowjobtype import WorkflowJobType
 from datetime import datetime
 from pydantic import model_serializer
-from typing_extensions import TypedDict
-from unstructured_client.types import BaseModel, Nullable, UNSET_SENTINEL
+from typing import List, Optional
+from typing_extensions import NotRequired, TypedDict
+from unstructured_client.types import (
+    BaseModel,
+    Nullable,
+    OptionalNullable,
+    UNSET,
+    UNSET_SENTINEL,
+)
 
 
 class JobInformationTypedDict(TypedDict):
     created_at: datetime
     id: str
-    runtime: Nullable[str]
     status: JobStatus
     workflow_id: str
     workflow_name: str
+    input_file_ids: NotRequired[Nullable[List[str]]]
+    job_type: NotRequired[WorkflowJobType]
+    output_node_files: NotRequired[Nullable[List[NodeFileMetadataTypedDict]]]
+    runtime: NotRequired[Nullable[str]]
 
 
 class JobInformation(BaseModel):
@@ -22,25 +34,31 @@ class JobInformation(BaseModel):
 
     id: str
 
-    runtime: Nullable[str]
-
     status: JobStatus
 
     workflow_id: str
 
     workflow_name: str
 
+    input_file_ids: OptionalNullable[List[str]] = UNSET
+
+    job_type: Optional[WorkflowJobType] = None
+
+    output_node_files: OptionalNullable[List[NodeFileMetadata]] = UNSET
+
+    runtime: OptionalNullable[str] = UNSET
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
-        optional_fields = []
-        nullable_fields = ["runtime"]
+        optional_fields = ["input_file_ids", "job_type", "output_node_files", "runtime"]
+        nullable_fields = ["input_file_ids", "output_node_files", "runtime"]
         null_default_fields = []
 
         serialized = handler(self)
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)