+++ src/unstructured_client/models/shared/sourceconnectortype.py
@@ -22,4 +22,7 @@ class SourceConnectorType(str, Enum):
     S3 = "s3"
     SALESFORCE = "salesforce"
     SHAREPOINT = "sharepoint"
+    SLACK = "slack"
     SNOWFLAKE = "snowflake"
+    JIRA = "jira"
+    ZENDESK = "zendesk"