+++ src/unstructured_client/models/shared/updatedestinationconnector.py
@@ -33,6 +33,10 @@ from .gcsdestinationconnectorconfiginput import (
     GCSDestinationConnectorConfigInput,
     GCSDestinationConnectorConfigInputTypedDict,
 )
+from .ibmwatsonxs3destinationconnectorconfiginput import (
+    IBMWatsonxS3DestinationConnectorConfigInput,
+    IBMWatsonxS3DestinationConnectorConfigInputTypedDict,
+)
 from .kafkaclouddestinationconnectorconfiginput import (
     KafkaCloudDestinationConnectorConfigInput,
     KafkaCloudDestinationConnectorConfigInputTypedDict,
@@ -52,6 +56,10 @@ from .neo4jdestinationconnectorconfiginput import (
 from .onedrivedestinationconnectorconfiginput import (
     OneDriveDestinationConnectorConfigInput,
     OneDriveDestinationConnectorConfigInputTypedDict,
+)
+from .pineconedestinationconnectorconfiginput import (
+    PineconeDestinationConnectorConfigInput,
+    PineconeDestinationConnectorConfigInputTypedDict,
 )
 from .postgresdestinationconnectorconfiginput import (
     PostgresDestinationConnectorConfigInput,
@@ -86,24 +94,26 @@ UpdateDestinationConnectorConfigTypedDict = TypeAliasType(
     "UpdateDestinationConnectorConfigTypedDict",
     Union[
         GCSDestinationConnectorConfigInputTypedDict,
+        ElasticsearchConnectorConfigInputTypedDict,
         AzureAISearchConnectorConfigInputTypedDict,
         WeaviateDestinationConnectorConfigInputTypedDict,
-        ElasticsearchConnectorConfigInputTypedDict,
         MongoDBConnectorConfigInputTypedDict,
         DeltaTableConnectorConfigInputTypedDict,
         QdrantCloudDestinationConnectorConfigInputTypedDict,
+        PineconeDestinationConnectorConfigInputTypedDict,
         Neo4jDestinationConnectorConfigInputTypedDict,
-        AstraDBConnectorConfigInputTypedDict,
-        S3DestinationConnectorConfigInputTypedDict,
         OneDriveDestinationConnectorConfigInputTypedDict,
-        DatabricksVDTDestinationConnectorConfigInputTypedDict,
+        S3DestinationConnectorConfigInputTypedDict,
+        AstraDBConnectorConfigInputTypedDict,
         MilvusDestinationConnectorConfigInputTypedDict,
+        DatabricksVolumesConnectorConfigInputTypedDict,
         PostgresDestinationConnectorConfigInputTypedDict,
         KafkaCloudDestinationConnectorConfigInputTypedDict,
-        DatabricksVolumesConnectorConfigInputTypedDict,
         CouchbaseDestinationConnectorConfigInputTypedDict,
         RedisDestinationConnectorConfigInputTypedDict,
+        DatabricksVDTDestinationConnectorConfigInputTypedDict,
         SnowflakeDestinationConnectorConfigInputTypedDict,
+        IBMWatsonxS3DestinationConnectorConfigInputTypedDict,
     ],
 )
 
@@ -112,24 +122,26 @@ UpdateDestinationConnectorConfig = TypeAliasType(
     "UpdateDestinationConnectorConfig",
     Union[
         GCSDestinationConnectorConfigInput,
+        ElasticsearchConnectorConfigInput,
         AzureAISearchConnectorConfigInput,
         WeaviateDestinationConnectorConfigInput,
-        ElasticsearchConnectorConfigInput,
         MongoDBConnectorConfigInput,
         DeltaTableConnectorConfigInput,
         QdrantCloudDestinationConnectorConfigInput,
+        PineconeDestinationConnectorConfigInput,
         Neo4jDestinationConnectorConfigInput,
-        AstraDBConnectorConfigInput,
-        S3DestinationConnectorConfigInput,
         OneDriveDestinationConnectorConfigInput,
-        DatabricksVDTDestinationConnectorConfigInput,
+        S3DestinationConnectorConfigInput,
+        AstraDBConnectorConfigInput,
         MilvusDestinationConnectorConfigInput,
+        DatabricksVolumesConnectorConfigInput,
         PostgresDestinationConnectorConfigInput,
         KafkaCloudDestinationConnectorConfigInput,
-        DatabricksVolumesConnectorConfigInput,
         CouchbaseDestinationConnectorConfigInput,
         RedisDestinationConnectorConfigInput,
+        DatabricksVDTDestinationConnectorConfigInput,
         SnowflakeDestinationConnectorConfigInput,
+        IBMWatsonxS3DestinationConnectorConfigInput,
     ],
 )