+++ src/unstructured_client/models/shared/updatesourceconnector.py
@@ -37,6 +37,10 @@ from .googledrivesourceconnectorconfiginput import (
     GoogleDriveSourceConnectorConfigInput,
     GoogleDriveSourceConnectorConfigInputTypedDict,
 )
+from .jirasourceconnectorconfiginput import (
+    JiraSourceConnectorConfigInput,
+    JiraSourceConnectorConfigInputTypedDict,
+)
 from .kafkacloudsourceconnectorconfiginput import (
     KafkaCloudSourceConnectorConfigInput,
     KafkaCloudSourceConnectorConfigInputTypedDict,
@@ -52,6 +56,10 @@ from .onedrivesourceconnectorconfiginput import (
 from .outlooksourceconnectorconfiginput import (
     OutlookSourceConnectorConfigInput,
     OutlookSourceConnectorConfigInputTypedDict,
+)
+from .postgressourceconnectorconfiginput import (
+    PostgresSourceConnectorConfigInput,
+    PostgresSourceConnectorConfigInputTypedDict,
 )
 from .s3sourceconnectorconfiginput import (
     S3SourceConnectorConfigInput,
@@ -69,6 +77,10 @@ from .snowflakesourceconnectorconfiginput import (
     SnowflakeSourceConnectorConfigInput,
     SnowflakeSourceConnectorConfigInputTypedDict,
 )
+from .zendesksourceconnectorconfiginput import (
+    ZendeskSourceConnectorConfigInput,
+    ZendeskSourceConnectorConfigInputTypedDict,
+)
 from typing import Union
 from typing_extensions import TypeAliasType, TypedDict
 from unstructured_client.types import BaseModel
@@ -79,19 +91,22 @@ UpdateSourceConnectorConfigTypedDict = TypeAliasType(
     Union[
         MongoDBConnectorConfigInputTypedDict,
         BoxSourceConnectorConfigInputTypedDict,
-        DropboxSourceConnectorConfigInputTypedDict,
-        ElasticsearchConnectorConfigInputTypedDict,
         GCSSourceConnectorConfigInputTypedDict,
-        SalesforceSourceConnectorConfigInputTypedDict,
+        ElasticsearchConnectorConfigInputTypedDict,
+        DropboxSourceConnectorConfigInputTypedDict,
         GoogleDriveSourceConnectorConfigInputTypedDict,
-        SharePointSourceConnectorConfigInputTypedDict,
+        SalesforceSourceConnectorConfigInputTypedDict,
+        ZendeskSourceConnectorConfigInputTypedDict,
         AzureSourceConnectorConfigInputTypedDict,
+        S3SourceConnectorConfigInputTypedDict,
         DatabricksVolumesConnectorConfigInputTypedDict,
         KafkaCloudSourceConnectorConfigInputTypedDict,
         OneDriveSourceConnectorConfigInputTypedDict,
         OutlookSourceConnectorConfigInputTypedDict,
-        S3SourceConnectorConfigInputTypedDict,
+        SharePointSourceConnectorConfigInputTypedDict,
         CouchbaseSourceConnectorConfigInputTypedDict,
+        PostgresSourceConnectorConfigInputTypedDict,
+        JiraSourceConnectorConfigInputTypedDict,
         ConfluenceSourceConnectorConfigInputTypedDict,
         SnowflakeSourceConnectorConfigInputTypedDict,
     ],
@@ -103,19 +118,22 @@ UpdateSourceConnectorConfig = TypeAliasType(
     Union[
         MongoDBConnectorConfigInput,
         BoxSourceConnectorConfigInput,
-        DropboxSourceConnectorConfigInput,
-        ElasticsearchConnectorConfigInput,
         GCSSourceConnectorConfigInput,
-        SalesforceSourceConnectorConfigInput,
+        ElasticsearchConnectorConfigInput,
+        DropboxSourceConnectorConfigInput,
         GoogleDriveSourceConnectorConfigInput,
-        SharePointSourceConnectorConfigInput,
+        SalesforceSourceConnectorConfigInput,
+        ZendeskSourceConnectorConfigInput,
         AzureSourceConnectorConfigInput,
+        S3SourceConnectorConfigInput,
         DatabricksVolumesConnectorConfigInput,
         KafkaCloudSourceConnectorConfigInput,
         OneDriveSourceConnectorConfigInput,
         OutlookSourceConnectorConfigInput,
-        S3SourceConnectorConfigInput,
+        SharePointSourceConnectorConfigInput,
         CouchbaseSourceConnectorConfigInput,
+        PostgresSourceConnectorConfigInput,
+        JiraSourceConnectorConfigInput,
         ConfluenceSourceConnectorConfigInput,
         SnowflakeSourceConnectorConfigInput,
     ],